requirements.txt
aenum==2.2.3
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
fbchat==1.9.7
idna==2.10
multidict==4.7.6
paho-mqtt==1.5.0
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
